[WIP] Things are broken here. Started refactoring page object.
Putting Overlay objects on the Page, rather than on the
top-level view component. Also, each page now needs to have
a content object -- the top-level component is used for less
now. (Should be just getting things from the URL and fetching
an object from the db.)

The Contents object should place the top bar on the page.

The Page object handles putting the CSS classes on the page.