Made the Last Workout link be a real, clickable link to the log
This involved adding a new query for logs that are associated
with unnamed workouts. Workouts now get an `identifier` added to
them when they are added to a program. Same idea as an _id, except
it's not an _id because they aren't top-level saveable objects.

The method `findLogsByWorkoutName` should be deleted and this new
`findLogsByWorkoutIdentifier` should always be used instead. That
way, workouts can go unnamed and still have logs associated with
them, and adding/changing the name later won't change which logs
are associated with it.