Made the title of an overlay get passed in as an attribute
Previously, the title was hard-coded in whatever file implemented
the overlay, and then exported from that file. Now it is passed in
as an attribute, so the title can be changed based on the contents
of the overlay window. E.g., an overlay showing an exercise can
have that exercise's name as the overlay title.