Return full workout logs, instead of just sets, for exercise history
This allows history to be grouped by the day it was performed,
instead of just a bunch of previous sets being shown. This will
be more logical and coherent for the user. They'll see "this is what
you did last workout," and can flip backwards by one workout at a
time.