Made the log date be a link to that actual log
*BUG* This has made me realize that there's a bug, which
has probably been here for a while. When clicking a link to
another page, which is of the same type as the current page,
it doesn't change.

This might be because of a similar issue that was causing
modal windows to not change when attributes updated, because
we are exporting a function that passes in the attrs for us,
rather than allowing the attrs to be a state of the page.

Try looking at changing the RenderPage function to not
pass in attrs for us.