Started removing exercise type and changing prescription format
Also removing confusing model types like SetLogVm,
GroupedSetLogVm, etc. Those shouldn't be necessary anymore
with the format of a prescription matching the format of a log.

The exercise model doesn't need to exist anymore -- they'll
only exist as part of a prescription. The unit type of the exercise
is determined by the prescribed units when that exercise appears
in workouts.